Fix pkg-config paths
Change shell_path helper function to always convert to forward slashes,
not just when cygpath is available. This is necessary as pkg_config_path
always converts to backslashes, and needs to be undone. This fixes
errors installing the generated .pc files due to 'bad escape sequences'
caused by paths having the wrong format.

(Note: as used above, "always" is short for "always (on Windows)"; none
of the described path munging occurs on other platforms.)